CyberScan.ai/backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
gunicorn>=21.2.0
sqlalchemy>=2.0.25
alembic>=1.13.1
asyncpg>=0.29.0
aiosqlite>=0.19.0
psycopg2-binary>=2.9.9
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
pydantic>=2.5.3
pydantic-settings>=2.1.0
dnspython>=2.4.2
httpx>=0.26.0
requests>=2.31.0
beautifulsoup4>=4.12.3
cryptography>=41.0.7
certifi>=2024.2.2
reportlab>=4.1.0
python-dotenv>=1.0.0
slowapi>=0.1.9
redis>=5.0.1
aiosmtplib>=3.0.1
email-validator>=2.1.0
apscheduler>=3.10.4
structlog>=24.1.0
pytest>=7.4.4
pytest-asyncio>=0.23.3
bcrypt>=4.0.0
